Integrate the readonly manager with the dump_manager.
